Struct pallet_lbp::pallet::Pallet

pub struct Pallet<T>(PhantomData<T>);

The pallet implementing the on-chain logic.

Tuple Fields
0: PhantomData<T>
Implementations

impl<T: Config> Pallet<T>

pub fn repay_fee() -> (u32, u32)

impl<T: Config> Pallet<T>

pub fn create_pool(
origin: OriginFor<T>,
pool_owner: T::AccountId,
asset_a: AssetId,
asset_a_amount: Balance,
asset_b: AssetId,
asset_b_amount: Balance,
initial_weight: LBPWeight,
final_weight: LBPWeight,
weight_curve: WeightCurveType,
fee: (u32, u32),
fee_collector: T::AccountId,
repay_target: Balance
) -> DispatchResult

Create a new liquidity bootstrapping pool for given asset pair.

For any asset pair, only one pool can exist at a time.

The dispatch origin for this call must be T::CreatePoolOrigin. The pool is created with initial liquidity provided by the pool_owner who must have sufficient funds free.

The pool starts uninitialized and update_pool call should be called once created to set the start block.

This function should be dispatched from governing entity T::CreatePoolOrigin

Parameters:

pool_owner: the future owner of the new pool.

asset_a: { asset_id, amount } Asset ID and initial liquidity amount.

asset_b: { asset_id, amount } Asset ID and initial liquidity amount.

initial_weight: Initial weight of the asset_a. 1_000_000 corresponding to 1% and 100_000_000 to 100% this should be higher than final weight

final_weight: Final weight of the asset_a. 1_000_000 corresponding to 1% and 100_000_000 to 100% this should be lower than initial weight

weight_curve: The weight function used to update the LBP weights. Currently, there is only one weight function implemented, the linear function.

fee: The trading fee charged on every trade distributed to fee_collector.

fee_collector: The account to which trading fees will be transferred.

repay_target: The amount of tokens to repay to separate fee_collector account. Until this amount is reached, fee will be increased to 20% and taken from the pool

Emits PoolCreated event when successful.

BEWARE: We are taking the fee from the accumulated asset. If the accumulated asset is sold to the pool, the fee cost is transferred to the pool. If its bought from the pool the buyer bears the cost. This increases the price of the sold asset on every trade. Make sure to only run this with previously illiquid assets.

pub fn update_pool_data(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId,
pool_owner: Option<T::AccountId>,
start: Option<T::BlockNumber>,
end: Option<T::BlockNumber>,
initial_weight: Option<LBPWeight>,
final_weight: Option<LBPWeight>,
fee: Option<(u32, u32)>,
fee_collector: Option<T::AccountId>,
repay_target: Option<Balance>
) -> DispatchResult

Update pool data of a pool.

The dispatch origin for this call must be signed by the pool owner.

The pool can be updated only if the sale has not already started.

At least one of the following optional parameters has to be specified.

Parameters:

pool_id: The identifier of the pool to be updated.

start: The new starting time of the sale. This parameter is optional.

end: The new ending time of the sale. This parameter is optional.

initial_weight: The new initial weight. This parameter is optional.

final_weight: The new final weight. This parameter is optional.

fee: The new trading fee charged on every trade. This parameter is optional.

fee_collector: The new receiver of trading fees. This parameter is optional.

Emits PoolUpdated event when successful.

pub fn add_liquidity(
origin: OriginFor<T>,
amount_a: (AssetId, <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance),
amount_b: (AssetId, <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance)
) -> DispatchResult

Add liquidity to a pool.

Assets to add has to match the pool assets. At least one amount has to be non-zero.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

pool_id: The identifier of the pool

amount_a: The identifier of the asset and the amount to add.

amount_b: The identifier of the second asset and the amount to add.

Emits LiquidityAdded event when successful.

pub fn remove_liquidity(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId
) -> DispatchResult

Transfer all the liquidity from a pool back to the pool owner and destroy the pool. The pool data are also removed from the storage.

The pool can’t be destroyed during the sale.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

amount_a: The identifier of the asset and the amount to add.

Emits ‘LiquidityRemoved’ when successful.

pub fn sell(
origin: OriginFor<T>,
asset_in: AssetId,
asset_out: AssetId,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance,
max_limit: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResult

Trade asset_in for asset_out.

Executes a swap of asset_in for asset_out. Price is determined by the pool and is affected by the amount and proportion of the pool assets and the weights.

Trading fee is distributed to the fee_collector.

Parameters:

asset_in: The identifier of the asset being transferred from the account to the pool.

asset_out: The identifier of the asset being transferred from the pool to the account.

amount: The amount of asset_in

max_limit: minimum amount of asset_out / amount of asset_out to be obtained from the pool in exchange for asset_in.

Emits SellExecuted when successful.

pub fn buy(
origin: OriginFor<T>,
asset_out: AssetId,
asset_in: AssetId,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance,
max_limit: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResult

Trade asset_in for asset_out.

Executes a swap of asset_in for asset_out. Price is determined by the pool and is affected by the amount and the proportion of the pool assets and the weights.

Trading fee is distributed to the fee_collector.

Parameters:

asset_in: The identifier of the asset being transferred from the account to the pool.

asset_out: The identifier of the asset being transferred from the pool to the account.

amount: The amount of asset_out.

max_limit: maximum amount of asset_in to be sold in exchange for asset_out.

Emits BuyExecuted when successful.

impl<T: Config> Pallet<T>

pub fn error_metadata() -> Option<PalletErrorMetadata>

impl<T: Config> Pallet<T>

pub(crate) fn deposit_event(event: Event<T>)

impl<T: Config> Pallet<T>

pub fn pool_data<KArg>(k: KArg) -> Option<Pool<T::AccountId, T::BlockNumber>> where
KArg: EncodeLike<<T as Config>::AccountId>,

Details of a pool.

impl<T: Config> Pallet<T>

pub(crate) fn calculate_weights(
pool_data: &Pool<T::AccountId, T::BlockNumber>,
at: T::BlockNumber
) -> Result<(LBPWeight, LBPWeight), DispatchError>

pub(crate) fn validate_pool_data(
pool_data: &Pool<T::AccountId, T::BlockNumber>
) -> DispatchResult

pub(crate) fn get_sorted_weight(
asset_in: AssetId,
now: T::BlockNumber,
pool_data: &Pool<T::AccountId, T::BlockNumber>
) -> Result<(LBPWeight, LBPWeight), Error<T>>

pub(crate) fn is_pool_running(
pool_data: &Pool<T::AccountId, T::BlockNumber>
) -> bool

return true if now is in interval <pool.start, pool.end>

pub(crate) fn has_pool_started(
pool_data: &Pool<T::AccountId, T::BlockNumber>
) -> bool

return true if now is > pool.start and pool has been initialized

pub(crate) fn collected_fees(
pool: &Pool<T::AccountId, T::BlockNumber>
) -> <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance

returns fees collected and locked in the fee collector account note: after LBP finishes and liquidity is removed this will be 0

pub(crate) fn is_repay_fee_applied(
pool: &Pool<T::AccountId, T::BlockNumber>
) -> bool

repay fee is applied until repay target amount is reached

pub(crate) fn execute_trade(
transfer: &AMMTransfer<T::AccountId, AssetId, AssetPair, Balance>
) -> DispatchResult

pub(crate) fn calculate_fees(
pool: &Pool<T::AccountId, T::BlockNumber>,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> Result<<<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance, DispatchError>

determines fee rate and applies it to the amount

pub fn pair_account_from_assets(
asset_a: AssetId,
asset_b: AssetId
) -> <T as Config>::AccountId