Struct pallet_lbp::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>

pub fn create_pool(
origin: OriginFor<T>,
pool_owner: T::AccountId,
asset_a: LBPAssetInfo<<<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance>,
asset_b: LBPAssetInfo<<<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance>,
sale_duration: (T::BlockNumber, T::BlockNumber),
weight_curve: WeightCurveType,
pausable: bool,
fee: Fee,
fee_receiver: T::AccountId
) -> DispatchResultWithPostInfo

Create a new liquidity bootstrapping pool for given asset pair.

For any asset pair, only one pool can exist at a time.

The dispatch origin for this call must be T::CreatePoolOrigin. The pool is created with initial liquidity provided by the pool_owner who must have sufficient funds free.

Parameters:

pool_owner: the owner of the new pool.

asset: The asset ID, the initial liquidity amount and the starting and ending weight.

sale_duration: The LBP event duration is determined by the starting and ending block number, or uninitialized if both values are set to zero.

weight_curve: The weight function used to update the LBP weights. Currently, there is only one weight function implemented, the linear function.

pausable: If the pausable option is set to true, the pool owner is allowed to pause the pool during the sale.

fee: The trading fee charged on every trade distributed to fee_receiver.

fee_receiver: The account to which trading fees will be transferred.

Emits PoolCreated event when successful.

pub fn update_pool_data(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId,
start: Option<T::BlockNumber>,
end: Option<T::BlockNumber>,
initial_weights: Option<((AssetId, LBPWeight), (AssetId, LBPWeight))>,
final_weights: Option<((AssetId, LBPWeight), (AssetId, LBPWeight))>,
fee: Option<Fee>,
fee_receiver: Option<T::AccountId>
) -> DispatchResultWithPostInfo

Update pool data of a pool.

The dispatch origin for this call must be signed by the pool owner.

The pool can be updated only if the sale has not already started.

At least one of the following optional parameters has to be specified.

Parameters:

pool_id: The identifier of the pool to be updated.

start: The new starting time of the sale. This parameter is optional.

end: The new ending time of the sale. This parameter is optional.

initial_weights: The new initial weights. This parameter is optional.

final_weights: The new final weights. This parameter is optional.

fee: The new trading fee charged on every trade. This parameter is optional.

fee_receiver: The new receiver of trading fees. This parameter is optional.

Emits PoolUpdated event when successful.

pub fn pause_pool(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId
) -> DispatchResultWithPostInfo

Pause a pool and disallow buy and sell operations on the pool.

Only a pool with the pausable option set to true can be paused.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

pool_id: The identifier of the pool

Emits Paused event when successful.

pub fn unpause_pool(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId
) -> DispatchResultWithPostInfo

Unpause a pool and allow token buy and sell operations on the pool.

A pool needs to be in the paused state prior this call.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

pool_id: The identifier of the pool

Emits Unpaused event when successful.

pub fn add_liquidity(
origin: OriginFor<T>,
amount_a: (AssetId, <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance),
amount_b: (AssetId, <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance)
) -> DispatchResultWithPostInfo

Add liquidity to a pool.

Assets to add has to match the pool assets. At least one amount has to be non-zero.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

pool_id: The identifier of the pool

amount_a: The identifier of the asset and the amount to add.

amount_b: The identifier of the second asset and the amount to add.

Emits LiquidityAdded event when successful.

pub fn remove_liquidity(
origin: OriginFor<T>,
pool_id: <T as Config>::AccountId
) -> DispatchResultWithPostInfo

Transfer all the liquidity from a pool back to the pool owner and destroy the pool. The pool data are also removed from the storage.

The pool can’t be destroyed during the sale.

The dispatch origin for this call must be signed by the pool owner.

Parameters:

amount_a: The identifier of the asset and the amount to add.

Emits ‘LiquidityRemoved’ when successful.

pub fn sell(
origin: OriginFor<T>,
asset_in: AssetId,
asset_out: AssetId,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance,
max_limit: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Trade asset_in for asset_out.

Executes a swap of asset_in for asset_out. Price is determined by the pool and is affected by the amount and proportion of the pool assets and the weights.

Trading fee is distributed to the fee_receiver.

Parameters:

asset_in: The identifier of the asset being transferred from the account to the pool.

asset_out: The identifier of the asset being transferred from the pool to the account.

amount: The amount of asset_in

max_limit: minimum amount of asset_out / amount of asset_out to be obtained from the pool in exchange for asset_in.

Emits SellExecuted when successful.

pub fn buy(
origin: OriginFor<T>,
asset_out: AssetId,
asset_in: AssetId,
amount: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance,
max_limit: <<T as Config>::MultiCurrency as MultiCurrency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Trade asset_in for asset_out.

Executes a swap of asset_in for asset_out. Price is determined by the pool and is affected by the amount and the proportion of the pool assets and the weights.

Trading fee is distributed to the fee_receiver.

Parameters:

asset_in: The identifier of the asset being transferred from the account to the pool.

asset_out: The identifier of the asset being transferred from the pool to the account.

amount: The amount of asset_out.

max_limit: maximum amount of asset_in to be sold in exchange for asset_out.

Emits BuyExecuted when successful.

impl<T: Config> Pallet<T>

pub fn pool_data<KArg>(k: KArg) -> Pool<T::AccountId, T::BlockNumber> where
KArg: EncodeLike<<T as Config>::AccountId>,

Details of a pool.